repetition checkers and better commentaries
Well, I just implemented a plentiful of functions to check if there is a repetition:
A) within the given file
B) within each subsequent move the player makes.

If there is a repetition, it will
in case a -> Tell the player where the repetition is, draw the board and exit with an error message saying 'Try Again'

in case b -> warn the player where the repetition is, and ask them gently to try again.