__name__ == "__main__", delete option and tip checker!
Well, first of all I added the function to check if a square in a position a player is trying to input is already filled.
If it is filled by a tip, the program won't let the player do so.
If it is filled by a previous move, the player's new move will be put in that position

also, there's the delete option!  Writing D<column>,<line> you can delete a move you made!

and, last but not least:
I added a main function, so I can use the __name__ == "__main__" , since I have a feeling i'm gonna need to use a second file to do the batch mode.