lots of stuff remade, game is now winnable and no commentaries
I just decided to leave the commentaries inside the code for when the program is done

I added the option to win, the game, and for this, i had to separate a first_tips list from the game_progression list.

The first_tips list is a list with all of the moves given by the initial tips file

the game_progression list is a list will all of the moves the player inputs

when you delete a move, this very move will be deleted in its respective list.

-I think the only thing left to do is the batch mode.